configure_datastore: Add an administrative user
It is necessary to add an administrative user to MongoDB, or else it will
not enforce authentication for local access.  Authentication is required
for access from a remote host, but not for any Unix user connecting from
the local host.

This means that if the data store is installed on a node host, users of the
PaaS will be able to ssh into their gears and access OpenShift's data store
without restriction.

Fortunately, the data store is not usually installed on a node host; such
a configuration is only likely in a POC setup.  However, we can still
provide better security for this unusual configuration.

This commit changes the kickstart script (and scripts generated from it) to
add an administrative user to MongoDB.  By default, this user has username
'admin' and password 'mongopass'.  These defaults can be changed by setting
values for CONF_MONGODB_ADMIN_USER and CONF_MONGODB_ADMIN_PASSWORD.

This commit also renames CONF_MONGODB_NAME to CONF_MONGODB_BROKER_USER, and
it renames CONF_MONGODB_PASSWORD to CONF_MONGODB_BROKER_PASSWORD.

CONF_MONGODB_*_{USER,PASSWORD} default to CONF_MONGODB_{USER,PASSWORD} if
the latter are set.

This change is related to BZ896395.