Switching mcollective to use the activemq plugin
The stomp plugin is technically deprecated.  The activemq plugin still uses the
stomp protocol.  The biggest difference I see is the different default
settings.  By default the activemq connector's retry algorithm uses an
exponential backoff with a maximum of 30 seconds.  Previously the stomp
connector would retry every 5 seconds if the connection was severed.