openshift.ks: set up infra repo for named
In configure_repos, define need_infra_repo to be true when installing
the named component.  This change is needed because the installation
script now installs the yum-plugin-priorities package, which is in the
openshift_infrastructure and openshift_node channels, and so at least
one of these two channels must be enabled.