openshift.sh: Parse command-line args
openshift.ks: Add a parse_cmdline function that parses the command-line
arguments.  Add a commented-out invocation of parse_cmdline.

scriptify, Makefile: When generating openshift.sh and openshift-amz.sh,
disable the invocation of parse_kernel_cmdline and enable the invocation of
parse_cmdline.

With this commit, it is now possible to specify settings as arguments
instead of environment variables.  For example, instead of the following:

    export CONF_DOMAIN=example.com
    sh openshift.sh

Now it will be possible to do the following:

    sh openshift.sh domain=example.com