<oo-admin-check-sources.py> Bug 1022825 - fix RHSM enable repo advice
This updates the output for RHSM to give correct instructions on how
to enable a currently disabled repo. This advice indicates the
subscription-manager tool, which will work right now and in upcoming
versions of the RHSM system (unlike the yum-config-manager tool, which
will only work for vanilla Yum repos in the future).

This commit also includes:

* fix erroneous reference to method calculate_enabled_repos(), which
  was removed in a previous revision.

* replaced _print_report method call with calls against logging()
  object logger

* cleaned out obsolete methods (_print_report, etc.) and member
  variables (replaced by repo_db module)