Update workflow for more sensible yum-plugin-priorities advice
oo-admin-check-sources.py now refrains from telling the user to
install yum-plugin-priorities until a repository is found that carries
the package. The package is provided by the OSE repos, so from a blank
system, the tool guides the user through the following steps:

* Initial run
** ```oo-admin-check-sources.py```
* Select a version of OpenShift
** ```oo-admin-check-sources.py -o 1.2```
* Select one or more system role(s)
** ```oo-admin-check-sources.py -o 1.2 -r node -r node-eap```
* Register a subscription that provides the needed repos
** ```subscription-manager attach --pool=xyz```
** ```oo-admin-check-sources.py -o 1.2 -r node -r node-eap```
* Enable the needed repos:
** ```subscription-manager repos --enable=repo_id```
** ```oo-admin-check-sources.py -o 1.2 -r node -r node-eap```
* Install yum-plugin-priorities
** ```yum install yum-plugin-priorities```
** ```oo-admin-check-sources.py -o 1.2 -r node -r node-eap```
* Set appropriate repository priorities
** ```oo-admin-check-sources.py -o 1.2 -r node -r node-eap --fix```

In order to perform the package search during the
yum-plugin-priorities verification step, ```_init_yumbase``` in
```check_sources.py``` needed to be updated to set a command line in
the ```YumBase.plugins``` object via ```setCmdLine```. This works
around a missing attribute error in the yum security plugin.