<oo-admin-check-sources.py> Bug 1026793, handle RepoError exceptions https://bugzilla.redhat.com/show_bug.cgi?id=1026793
Wrapped the `yum-plugin-priorities` check in try/except and added new
exception for "showstopper" errors. Most of the `RepoError` exceptions
that can bubble out of the yum API are unrecoverable.

There are probably other scenarios that can raise `RepoErrors` that
aren't being handled, but the changes to
`OpenShiftAdminCheckSources.main` allow us to gracefully exit for any
future showstoppers we might run across