<check-sources> Add rhel beta channel, update priorities code to handle multiple rhel channels
When using `beta2.ini`, the rhel server beta channel should be at the
same priority as the base rhel channel. This update adds the beta
channel to the ini file and updates `oo-admin-check-sources` to handle
multiple "rhel" product channels instead of just one.

This should be a lesson to me about specializing code