<install> prefix OpenShift: to output strategically
oo-install filters the output of openshift.sh so that a parallel
install of multiple hosts isn't just a confusing jumble of text.
The filter simply displays lines that begin with OpenShift:
So, this commit adds OpenShift in front of a few more lines that
help the user see that something is either happening or gone wrong.