<openshift.ks> node self-reports request for pub ssh key / bug 1054898
Updated 'wget' URL (used by the node to pull the broker's public key) to include a parameter containing the node hostname.  The idea is to allow the broker to accumulate knowledge of nodes configured in this way.  That information may be used down the line for streamlined registration of nodes with the broker system.

Luke Meyer: incorporated change into .sh and amz versions.
https://bugzilla.redhat.com/show_bug.cgi?id=1054898