openshift.ks: Fix logic in configure_rhn_channels
Fix a syntax error in configure_rhn_channels that would cause it to fail to
enable the infrastructure repository in the situation where we are
installing the host as neither a broker nor a node (i.e., if we are
installing only ActiveMQ, MongoDB, or BIND).

This commit fixes bug 1056895.