openshift.ks: Delete sed command for openshift.sh
openshift.ks: Delete the sed command that generates openshift.sh because
this sed command is no longer adequate (it is missing some step that the
Makefile performs to generate openshift.sh, and consequently it generates
a script that does not work properly).

README.md: Add a paragraph to explain the relationship between openshift.ks
and openshift.sh.