openshift.sh: Exit with 0 exit code
Make the last command in openshift.sh return true.

If we reach the last command in openshift.sh, then that means that nothing
called abort_install to terminate execution of openshift.sh, and so
openshift.sh should exit with a zero exit code irrespective of the success
of the previous command.