openshift.ks: CONF_ABORT_ON_UNRECOGNIZED_SETTINGS
Add the abort_on_unrecognized_settings/CONF_ABORT_ON_UNRECOGNIZED_SETTINGS
setting, which defaults to true and causes the installation script to abort
if an unrecognized setting is specified if enabled or causes the
installation script only to warn about unrecognized settings if disabled.

Make parse_kernel_cmdline set an overriding default value of
CONF_ABORT_ON_UNRECOGNIZED_SETTINGS=false so that kickstart scripts do not
abort because of kernel command-line arguments that are unrelated to the
installation script.

This commit fixes bug 1114519.