openshift.ks: Don't restart services for kickstart
Check whether the script is being run as a kickstart script and, if so,
do not restart services at the very end of the script.  Doing so is
unnecessary because the host is about to reboot anyway once the kickstart
script finishes its execution, it takes time, and it generates a bunch of
error messages for services that cannot successfully start in a kickstart
environment.

This commit fixes bug 1121293.