openshift.ks: Always set routing_plugin_user/pass
In set_defaults, set the routing_plugin_user and routing_plugin_pass
variables when installing the "broker" or "activemq" component instead of
setting them only when installing the "broker" component.  These variables
are used when configuring ActiveMQ and so may be used even if the "broker"
component is not being configured

This commit fixes bug 1110680.