admin scripts: script to run conn hooks for mongodb carts
This commit adds run-mongodb-connection-hooks.rb, intended to be run on
the broker, which automates running the connection hooks for any app
using the mongodb-2.4 add-on cartridge. This is specifically to update
the `PATH` and `LD_LIBRARY_PATH` variables for scaled carts after
updating to a new version of the `openshift-origin-cartridge-mongodb`
package.

Enterprise bug 1135062
https://bugzilla.redhat.com/show_bug.cgi?id=1135062