Add facter to the package exclusion list for 2.2 node repos
Puppet versions 3.x require newer versions of facter than those in the
node repos. With priorities in place this was preventing anyone from
upgrading or installing puppet 3.x. Recent versions of OpenShift
Enterprise only use the ruby193-facter package so excluding this
package from the repo is safe to do.