openshift.ks: Configure router and routing-daemon
Add router_hostname / CONF_ROUTER_HOSTNAME, router / CONF_ROUTER, and
enable_ha / CONF_ENABLE_HA settings, and add "router" as a component
recognized for the install_components / CONF_INSTALL_COMPONENTS setting.

Note: install_components / CONF_INSTALL_COMPONENTS defaults to all
components except the router.

Change configure_repos to enable the infrastructure and Software
Collections channels if installing the "router" component.

Change configure_hosts_dns to add a DNS record for the router.

Change install_broker_pkgs to install the routing daemon if CONF_ROUTER is
set to nginx.

Change configure_controller to set ALLOW_HA_APPLICATIONS, MANAGE_HA_DNS,
and DEFAULT_ALLOW_HA to true and ROUTER_HOSTNAME to the hostname specified
by CONF_ROUTER_HOSTNAME if CONF_ENABLE_HA is enabled.

Add install_router_pkgs, which is a no-op if CONF_ROUTER is set to "f5" and
installs nginx and the routing daemon if CONF_ROUTER is set to "nginx.

Call install_router_pkgs from install_rpms if installing the "router"
component.

Add configure_routing_daemon, which is a no-op unless CONF_ROUTER is set to
"f5" and we are installing the broker, or CONF_ROUTER is set to "nginx" and
we are installing the router, and otherwise configures
/etc/openshift/routing-daemon.conf and (if CONF_ROUTER is set to nginx)
restarts the daemon.

Call configure_routing_daemon from configure_openshift if installing the
broker or router.  (In the case of F5, the routing daemon runs on the
OpenShift broker; in the case of nginx, the routing daemon runs on the
router alongside nginx.)

Add configure_router, which is a no-up unless CONF_ROUTER is set to "nginx"
(the script cannot configure F5 for the user), in which case the function
configures the firewall and SELinux appropriately and starts nginx.

Call configure_router from configure_openshift if we are installing the
router.

This commit fixes bug 1223554.