openshift.ks: Rewrite set_conf() in sed
Rewrite set_conf() to use a single sed command and to insert lines more
sensibly.

Now if set_conf find an existing setting with a value other than the one
specified, it comments that setting out as before and inserts the new
setting after the old setting.  Otherwise, if the setting is not found in
the file, set_conf appends the new setting to the end of the file.