Properly detect role for node-eap/node-eap-6.3
This change introduces a fix for the `node-eap` role detection, which
was made wrinkly due to the need for a `node-eap-6.3` role. The new
`node-eap` detection criteria are:

* If the `openshift-origin-cartridge-jbosseap` package is installed,
  either the role `node-eap` or `node-eap-6.3` is needed
* If the EAP 6.3-specific channel is already enabled, assume the role
  `node-eap-6.3` is needed
* Otherwise, assume the `node-eap` role is needed, which entails
  enabling the EAP channel(s) bearing more current software versions

If the user specifies both the `node-eap` and `node-eap-6.3` roles,
assume they need the EAP 6.3-specific channel and use the
`node-eap-6.3` role accordingly.

In addition, the Enterprise bug 1098945 is *somewhat* addressed, in
that channels which:

* are "blessed", i.e. which appear in `/etc/repos.ini`
* but aren't required by detected/specified roles

are left `enabled` but have their `priority` set to the value for
`OTHER_PRIORITY` defined in `oo-admin-yum-validator`, which is
`40`. This prevents them from interfering with the required channels
as determined by the role(s).

This was introduced to handle the case where a user has both the
regular JBoss EAP channel enabled as well as the JBoss
EAP-6.3-specific channel. Having both channels enabled and at the same
priority would lead to a situation where packages from the regular EAP
channel would eventually shadow the EAP-6.3 packages as "upgrades",
which might not be what the user wants. Setting the conflicting
channel to the lower (higher numerically) priority prevents this from
happening.

Enterprise bug: 1098945
https://bugzilla.redhat.com/show_bug.cgi?id=1098945