More sensible node-eap special case logic
NOW `oo-admin-yum-validator` does its best to auto-detect the correct
`node-eap/-6.3` role based on:

* Which repo is enabled
* Which of the conflicting repo has higher priority
* Which is the "safe choice"

In that final case, if both 6.3 and "current" repos have the same
priority, oo-admin-yum-validator will prefer the 6.3 repo. This might
not be what the user intends, but it has less potential for harm than
the opposite case.

Also, instead of failing to `node-eap-6.3` if the user specifies both
`--role=node-eap` and `--role=node-eap-6.3`, `oo-admin-yum-validator`
now bails with an error message.