openshift.ks: Comments on word-splitting
Add explicit comments to indicate when we want word-splitting.

The one exception is when we are substituting a space for a separator in
a for loop, such as the following: for foo in ${bar//, / }

Other than instances of this exception, either variables should be
quoted or explicitly identified as needing word-splitting.

This commit is related to bug 1223551.