openshift.ks: Fix router configuration
Set the correct zone ($domain) when creating the router's DNS record.

Return 0 from configure_routing_daemon.  Before this commit, if $router
was set to nginx, configure_hosts_dns would return an error, and so
openshift.ks/openshift.sh would terminate because we use set -e now.

In set_defaults, abort if the router component is enabled at the same time
as the broker or node components.  Because all of these components listen
on ports 80 and 443, the router cannot co-exist with the broker or node
without additional configuration, and it does not make a lot of sense to
colocate the router with the other components anyway.