Add smarter config file handling, etc.
This adds smarter config file saving and reading.

Values which should ride with the config settings, but shouldn't be
saved, are now filtered prior to saving.

Unattended mode is settable with a flag. New options need to check this
before going into prompt-the-user mode.

The Ansible config file can be specified but isn't used yet.
* certain ansible config options need to be derived and saved to the
  actual config location
* the config needs to be copied to one of these places:
  * ANSIBLE_CONFIG (an environment variable)
  * ansible.cfg (in the current directory)
  * .ansible.cfg (in the home directory)
  * /etc/ansible/ansible.cfg

Other bits and pieces.