Ensure pre-existing gem path is still taken into consideration for gems not provided by oo-install
Previously, any defined `$GEM_PATH` variable would not be taken into consideration. If gems were installed in a non-default location, it is possible oo-install would not find gems it does not provide but expects to be installed. This change will cause oo-install to search its own list of gems for a gem first, then search the previously defined `$GEM_PATH` for the gem.

Additionally, when compiling the `$RUBYLIB` variable by looping through `$OOINSTALL_GEM_PATH`, it was possible before for some entries in the loop to not end
with '/', resulting in incorrect paths added for gems. For example:

```
/tmp/oo-install-ose-20140908-1720/lib:/tmp/oo-install/vendor/bundle:/tmp/oo-install-ose-20140908-1720/vendor/bundle/ruby/gemsdiff-lcs-1.2.4/lib/:/tm...
```

This issue has been resolved by appending a `/` on the path added to RUBYLIB where neccessary.

The `$RUBYLIB` variable is now compiled using both the paths from the original `$GEM_PATH` and the `OOINSTALL_GEM_PATH`

Bug 1223324
https://bugzilla.redhat.com/show_bug.cgi?id=1223324